worked on landmakr to itineary view

+++ etc/LandmarkToursScript.sql
@@ -49,8 +49,8 @@ id INTEGER IDENTITY(1,1) not null,
 name VARCHAR(64) not null,
 itinerary_DATE DATE not null,
 user_id INTEGER not null,
-starting_latitude INTEGER not null,
-starting_longitude INTEGER not null
+starting_latitude FLOAT not null,
+starting_longitude FLOAT not null
 
 CONSTRAINT pk_itinerary_id PRIMARY KEY (id),
 CONSTRAINT fk_itinerary_user_id FOREIGN KEY (user_id) REFERENCES app_user(id)
